+++ R/utils.R
@@ -1,3 +1,17 @@
+#' Shared Parameters
+#'
+#' @description Contains arguments that are shared between multiple functions
+#'   in the package to avoid repetition using \code{inheritParams}.
+#'
+#' @param plot_height optional, (\code{numeric}) a vector of length three with \code{c(value, min, max)}. Specifies
+#'   the height of the main plot.
+#' @param plot_width optional, (\code{numeric}) a vector of length three with \code{c(value, min, max)}. Specifies
+#'   the width of the main plot and renders a slider on the plot to interactively adjust the plot width.
+#' @param label (\code{character}) module label in the teal app. Please note that this module is developed based on
+#' ADaM data structure and ADaM variables.
+#'
+#' @name shared_params
+NULL
 
 #'
 #' Utility function for quick filter