+++ NEWS.md
@@ -1,5 +1,6 @@
 # teal.osprey 0.1.6
 
+* All graph modules now accept a `plot_width` argument which specifies the plot width and renders a slider to adjust the width interactively in the module.
 * Replaced `plot_with_height` with `plot_with_settings` module throughout package.
 
 # teal.osprey 0.1.5